moved domain logic out of the JsonReader
