+++ src/main/java/amidst/mojangapi/file/service/DotMinecraftDirectoryService.java
@@ -9,11 +9,11 @@ import amidst.documentation.Immutable;
 import amidst.documentation.NotNull;
 import amidst.logging.AmidstLogger;
 import amidst.mojangapi.MojangApi;
-import amidst.mojangapi.file.JsonReader;
 import amidst.mojangapi.file.MojangApiParsingException;
 import amidst.mojangapi.file.directory.DotMinecraftDirectory;
 import amidst.mojangapi.file.directory.ProfileDirectory;
 import amidst.mojangapi.file.directory.VersionDirectory;
+import amidst.mojangapi.file.json.JsonReader;
 import amidst.mojangapi.file.json.ReleaseType;
 import amidst.mojangapi.file.json.launcherprofiles.LauncherProfileJson;
 import amidst.mojangapi.file.json.launcherprofiles.LauncherProfilesJson;
@@ -55,7 +55,7 @@ public class DotMinecraftDirectoryService {
 	public LauncherProfilesJson readLauncherProfilesFrom(DotMinecraftDirectory dotMinecraftDirectory)
 			throws MojangApiParsingException,
 			IOException {
-		return JsonReader.readLauncherProfilesFrom(dotMinecraftDirectory.getLauncherProfilesJson());
+		return JsonReader.readLocation(dotMinecraftDirectory.getLauncherProfilesJson(), LauncherProfilesJson.class);
 	}
 
 	@NotNull
